support adding roles to teams
