+++ frontend/controller/access/teams/TeamPage/TeamDetails.tsx
@@ -1,0 +1,55 @@
+/* eslint-disable @typescript-eslint/no-non-null-assertion */
+import { useTranslation } from 'react-i18next';
+import { useNavigate } from 'react-router-dom';
+import { PageDetail, PageDetails, SinceCell, TextCell } from '../../../../../framework';
+import { RouteE } from '../../../../Routes';
+import { Team } from '../../../interfaces/Team';
+
+export function TeamDetails(props: { team: Team }) {
+  const { t } = useTranslation();
+  const { team } = props;
+  const history = useNavigate();
+  return (
+    <PageDetails>
+      <PageDetail label={t('Name')}>{team.name}</PageDetail>
+      <PageDetail label={t('Description')}>{team.description}</PageDetail>
+      <PageDetail label={t('Organization')}>
+        <TextCell
+          text={team.summary_fields?.organization?.name}
+          to={RouteE.OrganizationDetails.replace(
+            ':id',
+            (team.summary_fields?.organization?.id ?? '').toString()
+          )}
+        />
+      </PageDetail>
+      <PageDetail label={t('Created')}>
+        <SinceCell
+          value={team.created}
+          author={team.summary_fields?.created_by?.username}
+          onClick={() =>
+            history(
+              RouteE.UserDetails.replace(
+                ':id',
+                (team.summary_fields?.created_by?.id ?? 0).toString()
+              )
+            )
+          }
+        />
+      </PageDetail>
+      <PageDetail label={t('Last modified')}>
+        <SinceCell
+          value={team.modified}
+          author={team.summary_fields?.modified_by?.username}
+          onClick={() =>
+            history(
+              RouteE.UserDetails.replace(
+                ':id',
+                (team.summary_fields?.modified_by?.id ?? 0).toString()
+              )
+            )
+          }
+        />
+      </PageDetail>
+    </PageDetails>
+  );
+}
