+++ frontend/controller/access/users/UserPage/AddRolesToUser.tsx
@@ -1,10 +1,10 @@
 import { useTranslation } from 'react-i18next';
 import { useNavigate, useParams } from 'react-router-dom';
-import { PageHeader, PageLayout } from '../../../../framework';
-import { LoadingPage } from '../../../../framework/components/LoadingPage';
-import { useItem } from '../../../common/useItem';
-import { User } from '../../interfaces/User';
-import { AddRolesToUsersForm } from '../roles/AddRolesToUsersForm';
+import { PageHeader, PageLayout } from '../../../../../framework';
+import { LoadingPage } from '../../../../../framework/components/LoadingPage';
+import { useItem } from '../../../../common/useItem';
+import { User } from '../../../interfaces/User';
+import { AddRolesForm } from '../../roles/AddRolesForm';
 
 export function AddRolesToUser() {
   const { t } = useTranslation();
@@ -15,7 +15,7 @@ export function AddRolesToUser() {
   return (
     <PageLayout>
       <PageHeader title={t('Add roles to user')} />
-      <AddRolesToUsersForm users={[user]} onClose={() => navigate(-1)} />
+      <AddRolesForm users={[user]} onClose={() => navigate(-1)} />
     </PageLayout>
   );
 }
